feat：优化演示与接口出参

+++ wemirr-platform-authority/src/main/resources/db/tenant_schema.sql
@@ -323,7 +323,7 @@ COMMIT;
 DROP TABLE IF EXISTS `sys_role`;
 CREATE TABLE `sys_role` (
                             `id` bigint NOT NULL AUTO_INCREMENT,
-                            `tenant_id` int DEFAULT NULL COMMENT '租户编码',
+                            `tenant_id` bigint DEFAULT NULL COMMENT '租户编码',
                             `code` varchar(30) DEFAULT NULL COMMENT '角色编码',
                             `name` varchar(30) NOT NULL DEFAULT '' COMMENT '名称',
                             `description` varchar(255) DEFAULT NULL COMMENT '描述信息',
